fix: 문제 수 검증 로직 변경

+++ server/src/main/java/com/pullup/exam/service/ExamService.java
@@ -263,8 +263,11 @@ public class ExamService {
         examProblemRepository.saveAll(examProblems);
     }
 
-    private void validateProblemCount(List<Subject> subjects, CorrectRateRange correctRateRange,
-                                      Long minValidSubejectCount) {
+    private void validateProblemCount(
+            List<Subject> subjects,
+            CorrectRateRange correctRateRange,
+            int minValidSubejectCount
+    ) {
         for (Subject subject : subjects) {
             long problemCount = problemRepository.countBySubjectAndCorrectRateBetween(
                     subject,
